move visitor to series_visitor; implement, chsnge to post order
