+++ symengine/visitor.h
@@ -55,6 +55,7 @@ public:
 };
 
 void preorder_traversal_stop(const Basic &b, StopVisitor &v);
+void postorder_traversal_stop(const Basic &b, StopVisitor &v);
 
 class HasSymbolVisitor : public BaseVisitor<HasSymbolVisitor, StopVisitor> {
 protected:
@@ -145,65 +146,6 @@ RCP<const Basic> coeff(const Basic &b, const RCP<const Basic> &x,
         const RCP<const Basic> &n);
 
 set_basic free_symbols(const Basic &b);
-
-class NeedsSymbolicExpansionVisitor : public BaseVisitor<NeedsSymbolicExpansionVisitor, StopVisitor> {
-protected:
-    RCP<const Symbol> x_;
-    bool needs_;
-public:
-
-    void bvisit(const TrigFunction &f) {
-        auto arg = f.get_arg();
-        map_basic_basic subsx0{{x_, integer(0)}};
-        if (arg->subs(subsx0)->__neq__(*integer(0))) {
-            needs_ = true;
-            stop_ = true;
-        }
-    }
-
-    void bvisit(const HyperbolicFunction &f) {
-        auto arg = f.get_arg();
-        map_basic_basic subsx0{{x_, integer(0)}};
-        if (arg->subs(subsx0)->__neq__(*integer(0))) {
-            needs_ = true;
-            stop_ = true;
-        }
-    }
-
-    void bvisit(const Pow &pow) {
-        auto base = pow.get_base();
-        auto exp = pow.get_exp();
-        map_basic_basic subsx0{{x_, integer(0)}};
-        // exp(const) or x^-1
-        if ((base->__eq__(*E) and exp->subs(subsx0)->__neq__(*integer(0)))
-            or (is_a_Number(*exp) and static_cast<const Number&>(*exp).is_negative()
-                and base->subs(subsx0)->__eq__(*integer(0)))) {
-            needs_ = true;
-            stop_ = true;
-        }
-    }
-
-    void bvisit(const Log &f) {
-        auto arg = f.get_arg();
-        map_basic_basic subsx0{{x_, integer(0)}};
-        if (arg->subs(subsx0)->__eq__(*integer(0))) {
-            needs_ = true;
-            stop_ = true;
-        }
-    }
-
-    void bvisit(const LambertW &x) { needs_ = true; stop_ = true; }
-
-    void bvisit(const Basic &x) { }
-
-    bool apply(const Basic &b, const RCP<const Symbol> &x) {
-        x_ = x;
-        needs_ = false;
-        stop_ = false;
-        preorder_traversal_stop(b, *this);
-        return needs_;
-    }
-};
 
 } // SymEngine
